page file Refactoring
