+++ spec/pages/dashboard/dashboard_page.rb
