+++ spec/pages/patients/appointments_page.rb
@@ -1,0 +1,3 @@
+class AppointmentsPage < ApplicationPage
+
+end
