+++ spec/pages/protocol_drugs/protocol_detail_page.rb
